requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.2
spacy>=3.0.0
python-multipart>=0.0.5
python-dotenv>=0.19.0
ollama>=0.0.6